+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Types of changes include:
 ## Upcoming/Unreleased
 
 - NEW: Target-based goals now show fun stats about your progress and encouragment. (h/t einhornlasagne)
+- CHANGED: The links for the About, Privacy, Changelog, and Support pages are now located under the user icon in the top right.
 - FIXED: When navigating from project to project, project graphs would sometimes not display. That's now fixed.
 
 ## 0.9.2